(update) Breaks out extension classes into individual files. Removes unused namespaces.
