+++ src/PhoenixSystem.Engine/Game/IGameManager.cs
@@ -11,14 +11,11 @@ namespace PhoenixSystem.Engine.Game
         IEntityManager EntityManager { get; }
         IEnumerable<IManager> Managers { get; }
         IEnumerable<ISystem> Systems { get; }
-
         IEnumerable<IDrawableSystem> DrawableSystems { get; }
         bool IsUpdating { get; }
-        void Update(ITickEvent tickEvent);
-
         bool IsDrawing { get; }
+        void Update(ITickEvent tickEvent);
         void Draw(ITickEvent tickEvent);
-
         event EventHandler EntityAdded;
         void AddEntity(IEntity e);
         void AddEntities(IEnumerable<IEntity> entities);
