+++ src/PhoenixSystem.Engine/IDrawable.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PhoenixSystem.Engine
+﻿namespace PhoenixSystem.Engine
 {
     public interface IDrawable
     {
