+++ src/PhoenixSystem.Engine/IManager.cs
@@ -1,12 +1,11 @@
-﻿using PhoenixSystem.Engine.Channel;
+﻿using System;
+using PhoenixSystem.Engine.Channel;
 using PhoenixSystem.Engine.Game;
-using System;
 
 namespace PhoenixSystem.Engine
 {
     public interface IManager : IChannelFilterable, IComparable<IManager>
     {
-        
         int Priority { get; }
         void Update();
         void Register(IGameManager gameManager);
