+++ src/PhoenixSystem.Engine/System/IDrawableSystem.cs
@@ -1,15 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PhoenixSystem.Engine.System
+﻿namespace PhoenixSystem.Engine.System
 {
     /// <summary>
-    /// System with support for Rendering stuff.
+    ///     System with support for Rendering stuff.
     /// </summary>
-    public interface IDrawableSystem :ISystem, IDrawable
+    public interface IDrawableSystem : ISystem, IDrawable
     {
     }
 }
